move content safety shared files to common package
