--- src/promptflow-evals/promptflow/evals/evaluators/_content_safety/common/__init__.py
@@ -1,7 +1,0 @@
-# ---------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# ---------------------------------------------------------
-
-from .content_safety_base import ContentSafetyEvaluatorBase
-
-__all__ = ["ContentSafetyEvaluatorBase"]
